optimized zoom in map builder state

+++ SP Project/MapBuilderState.h
@@ -28,7 +28,7 @@ private:
 	tex_tile picked_tile = { 0,0 };
 	int size_x, size_y;  //<-- map size
 	int x = 0, y = 0;    //<-- location of upper left corner of the map
-	int x_offset = 0, y_offset = 0; //<-- offset from upper left corner of the screen to upper left corner of the map
+	float x_offset = 0, y_offset = 0; //<-- offset from upper left corner of the screen to upper left corner of the map
 	int x_mid = 0, y_mid = 0;       //<-- index of the tile in the center of the screen
 	int cam_speed = 150;
 	bool active_grid = 1; //<-- grid active/inactive
